chore: move CreateSampleComponent to separate partials
