+++ src/Uno.UI.Runtime.Skia.Wpf/Extensions/WpfNativeElementHostingExtension.cs
@@ -1,9 +1,7 @@
 ﻿#nullable enable
 
 using System;
-using System.Windows.Controls;
 using System.Windows.Media;
-using System.Windows.Shapes;
 using Uno.Foundation.Logging;
 using Microsoft.UI.Xaml;
 using ContentPresenter = Microsoft.UI.Xaml.Controls.ContentPresenter;
@@ -11,7 +9,7 @@ using WpfCanvas = System.Windows.Controls.Canvas;
 
 namespace Uno.UI.Runtime.Skia.Wpf;
 
-internal class WpfNativeElementHostingExtension : ContentPresenter.INativeElementHostingExtension
+internal partial class WpfNativeElementHostingExtension : ContentPresenter.INativeElementHostingExtension
 {
 	private readonly ContentPresenter _presenter;
 
@@ -107,34 +105,5 @@ internal class WpfNativeElementHostingExtension : ContentPresenter.INativeElemen
 		}
 
 		return Windows.Foundation.Size.Empty;
-	}
-
-	public object CreateSampleComponent(string text)
-	{
-		return new Button
-		{
-			Width = 100,
-			Height = 100,
-			Content = new Viewbox
-			{
-				Child = new StackPanel
-				{
-					Children =
-					{
-						new TextBlock
-						{
-							Text = text
-						},
-						new Path
-						{
-							// A star
-							Data = Geometry.Parse("M 17.416,32.25L 32.910,32.25L 38,18L 43.089,32.25L 58.583,32.25L 45.679,41.494L 51.458,56L 38,48.083L 26.125,56L 30.597,41.710L 17.416,32.25 Z"),
-							Stretch = Stretch.Uniform,
-							Stroke = Brushes.Red
-						}
-					}
-				}
-			}
-		};
 	}
 }
